Refactors category matching configuration
Updates category matching configuration to use a HashSet for parameters, improving performance and simplifying merging logic.

This change introduces a `MatcherValue` record to encapsulate value/amount pairs, and uses this in conjunction with HashSet to streamline deduplication and merging of matchers.

The configuration file `categories.json` is added, and old version deleted.

The change also removes case sensitivity for regular expressions.